Update and rename V_1_0_3__CREATE TABLE hero.sql to V003__CREATE TABLE hero.sql

+++ V003__CREATE TABLE hero.sql
@@ -1,5 +1,5 @@
 CREATE TABLE hero(
-    id BIGSERIAL NOT NULL,
+    id  INT NOT NULL,
     name VARCHAR(250) NOT NULL,
     description TEXT NOT NULL,
     debut_year INT NOT NULL,
